requirements.txt
pipeline-ai # mystic.ai
cog # replicate
replicate # replicate
modal # modal
libsql-client # To save benchmarks
requests # Modal run benchmarks
python-dotenv # Modal run benchmarks
inferless-cli # inferless
keyrings.alt # inferless
runpod # runpod
beam-sdk # beam
